Update TRANSLATORS file.
The header of that file already stated that only current translators were
listed in that file. So there is no need to mark the current translators
with a star, I just removed the old ones instead (all history of that file
is kept in git anyway).

Current translators = all translators who contributed to 3.1.x translations.

Signed-off-by: Chantry Xavier <shiningxc@gmail.com>