Add missing NULL set in group printing
If we don't set the pkgname var to NULL, we run into all sorts of beautiful
segfault behavior when a group spans multiple repositories and we try to
print out the location of the former list. Easy fix.

This regression was introduced in bf8670036907b3ede18d37b7a3f0f7e14542a5ac.

Signed-off-by: Dan McGee <dan@archlinux.org>